Rag-agent(bot)/requirements.txt
python-dotenv==1.0.1
qdrant-client==1.9.2
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
langchain-core==0.2.27
langchain-qdrant==0.1.3
tqdm==4.66.5
pydantic==2.8.2
rapidfuzz==3.9.6
pandas==2.2.2
openpyxl==3.1.2
langgraph==0.1.19
